Using getter for idTable
- This way this change does not break anything
- Added a test case for testing this fix.
- Removed demonstration files
