+++ jwz.js
@@ -112,8 +112,6 @@
     return function() {
       var idTable = {};
 
-      function getIdTable() { return idTable; }
-      
       function thread(messages) {
         idTable = this.createIdTable(messages);
         var root = messageContainer();
@@ -271,7 +269,7 @@
         pruneEmpties: pruneEmpties,
         groupBySubject: groupBySubject,
         thread: thread,
-        getIdTable: getIdTable
+        get idTable() { return idTable; }
       }
     }();
   }
